Move all of the google realtime stuff out of the canvas elements and CanvasService
