+++ src/app/canvas/canvas/canvas.component.html
@@ -5,12 +5,12 @@
      (mousedown)="this.onMousedown($event)"
      (mousewheel)="this.onMousewheel($event)">
   <g uxg-card
-     *ngFor="let card of (this.canvasService.cards | toIterable)"
+     *ngFor="let card of (this.canvasService.cards)"
      [card]="card"
      [attr.card-id]="card.id"
      [canvasBoundsGetter]="getBounds"></g>
   <g uxg-arrow
-     *ngFor="let arrow of (this.canvasService.arrows | toIterable)"
+     *ngFor="let arrow of (this.canvasService.arrows)"
      [arrow]="arrow"
      [canvasBoundsGetter]="getBounds"></g>
 </svg>
